+++ packages/webview_flutter/webview_flutter_platform_interface/CHANGELOG.md
@@ -4,6 +4,8 @@
   `PlatformWebViewController.setOnPlatformPermissionRequest`.
 * Updates minimum Flutter version to 3.3.
 * Fixes common typos in tests and documentation.
+* Adds support for listening to HTTP errors. See
+  `PlatformNavigationDelegate.setOnHttpError`.
 
 ## 2.1.0