+++ packages/webview_flutter/webview_flutter_wkwebview/CHANGELOG.md
@@ -1,6 +1,10 @@
-## 3.3.0
+## 3.4.0
 
 * Adds support for `PlatformWebViewController.setOnPlatformPermissionRequest`.
+
+## 3.3.0
+
+* Adds support for `PlatformNavigationDelegate.onUrlChange`.
 
 ## 3.2.4